python-service/requirements.txt
alembic==1.15.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-tools==1.71.0
Mako==1.3.9
MarkupSafe==3.0.2
protobuf==5.29.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
